Feature/swagger and test (#4)
* create swagger

* create test unitary and create docker-compose

+++ src/core/modules/delete-shorts-links-usecase.ts
@@ -13,9 +13,12 @@ export class DeleteShortsLinksUseCase implements IDeleteShortsLinkUseCase {
     this.shortsLinksRepo = shortsLinksRepo;
   }
 
-  async delete (code: string, userId: string): Promise<Result<any>> {
+  async delete (code: string): Promise<Result<any>> {
     const shortLinksExists = await this.shortsLinksRepo.findByUrlShorts(code);
+    if (!shortLinksExists) {
+      return Result.fail('URL shorts not exists.');
+    }
     await this.shortsLinksRepo.delete(shortLinksExists.id);
-    return Result.ok(userId);
+    return Result.ok({ id: shortLinksExists.id });
   }
 }
